+++ src/Command/FilterDependencyCollectionCommand.php
@@ -5,39 +5,19 @@ declare(strict_types=1);
 namespace ComposerUnused\ComposerUnused\Command;
 
 use ComposerUnused\ComposerUnused\Dependency\DependencyCollection;
-
-use function array_merge;
+use ComposerUnused\ComposerUnused\Filter\FilterCollection;
 
 final class FilterDependencyCollectionCommand
 {
-    private const GLOBAL_NAMED_EXCLUSION = [
-        'composer-plugin-api'
-    ];
+    private DependencyCollection $requiredDependencyCollection;
+    private FilterCollection $filterCollection;
 
-    private const GLOBAL_PATTERN_EXCLUSION = [
-        '/-implementation$/i'
-    ];
-
-    /** @var DependencyCollection */
-    private $requiredDependencyCollection;
-    /** @var array<string> */
-    private $namedExclusion;
-    /** @var array<string> */
-    private $patternExclusion;
-
-    /**
-     * @param DependencyCollection $requiredDependencyCollection
-     * @param array<string> $namedExclusion
-     * @param array<string> $patternExclusion
-     */
     public function __construct(
         DependencyCollection $requiredDependencyCollection,
-        array $namedExclusion = [],
-        array $patternExclusion = []
+        FilterCollection $filterCollection
     ) {
         $this->requiredDependencyCollection = $requiredDependencyCollection;
-        $this->namedExclusion = array_merge(self::GLOBAL_NAMED_EXCLUSION, $namedExclusion);
-        $this->patternExclusion = array_merge(self::GLOBAL_PATTERN_EXCLUSION, $patternExclusion);
+        $this->filterCollection = $filterCollection;
     }
 
     public function getRequiredDependencyCollection(): DependencyCollection
@@ -45,19 +25,8 @@ final class FilterDependencyCollectionCommand
         return $this->requiredDependencyCollection;
     }
 
-    /**
-     * @return array<string>
-     */
-    public function getNamedExclusion(): array
+    public function getFilters(): FilterCollection
     {
-        return $this->namedExclusion;
-    }
-
-    /**
-     * @return array<string>
-     */
-    public function getPatternExclusion(): array
-    {
-        return $this->patternExclusion;
+        return $this->filterCollection;
     }
 }