+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added CLI argument `composer-json` which can be used to parse external projects.
   This will default to the current working directory.
 - Added error message when `composer.json` is not readable
+- Added check for zombie exclusion. It will report if any excluded packages or pattern did match any package
 ### Changed
 - Change `bin/composer-unused` to be used as standalone binary
 - Package type is now `library` instead of `composer-plugin`